update LockTargetState to match LockCurrentState during lock refresh (#144)
## :recycle: Current situation

The LockTargetState attribute is not updated while polling/listening to
events from the August API. This leads to any manual changes of the lock
(ie I physically lock/unlock it) to cause the lock status to always show
unlocking or locking in Homekit.

## :bulb: Proposed solution

Update the LockTargetState attribute to be the same value as the
LockCurrentState during both polling or pushed events. This plugin used
to operate with this logic around the 2.0.0 version and prior but
somewhere after 2.0.0 this logic was changed.

## :gear: Release Notes

LockTargetState is updated for manual lock events

## :heavy_plus_sign: Additional Information


### Testing


### Reviewer Nudging

Co-authored-by: Donavan Becker <beckersmarthome@icloud.com>